+++ src-docs/src/views/steps/steps_example.js
@@ -9,6 +9,7 @@ import {
 import {
   EuiCode,
   EuiSteps,
+  EuiStep,
 } from '../../../../src/components';
 
 import Steps from './steps';
@@ -27,6 +28,10 @@ import StepsHorizontal from './steps_horizontal';
 const stepsHorizontalSource = require('!!raw-loader!./steps_horizontal');
 const stepsHorizontalHtml = renderToHtml(StepsHorizontal);
 
+import Status from './status';
+const statusSource = require('!!raw-loader!./status');
+const statusHtml = renderToHtml(Steps);
+
 export const StepsExample = {
   title: 'Steps',
   sections: [{
@@ -42,7 +47,7 @@ export const StepsExample = {
         Numbered steps
       </p>
     ),
-    props: { EuiSteps },
+    props: { EuiSteps, EuiStep },
     demo: <Steps />,
   },
   {
@@ -86,6 +91,24 @@ export const StepsExample = {
       </div>
     ),
     demo: <HeadingElementSteps />,
+  },
+  {
+    title: 'Steps status',
+    source: [{
+      type: GuideSectionTypes.JS,
+      code: statusSource,
+    }, {
+      type: GuideSectionTypes.HTML,
+      code: statusHtml,
+    }],
+    text: (
+      <p>
+        Steps can optionally include <EuiCode>status</EuiCode> prop with
+        a value of <EuiCode>complete</EuiCode> or <EuiCode>incomplete</EuiCode>. This
+        is used mostly as a final step when you need to make some sort of final check.
+      </p>
+    ),
+    demo: <Status />,
   },
   {
     title: 'Horizontal',