+++ src-docs/src/views/text/text_example.js
@@ -24,6 +24,10 @@ import TextColor from './text_color';
 const textColorSource = require('!!raw-loader!./text_color');
 const textColorHtml = renderToHtml(TextColor);
 
+import TextWidth from './text_width';
+const textWidthSource = require('!!raw-loader!./text_width');
+const textWidthHtml = renderToHtml(TextWidth);
+
 export const TextExample = {
   title: 'Text',
   sections: [{
@@ -67,6 +71,23 @@ export const TextExample = {
       </p>
     ),
     demo: <TextSmall />,
+  }, {
+    title: 'Changing width',
+    source: [{
+      type: GuideSectionTypes.JS,
+      code: textWidthSource,
+    }, {
+      type: GuideSectionTypes.HTML,
+      code: textWidthHtml,
+    }],
+    text: (
+      <p>
+        In situations where you need to allow the width to expand to fill its container,
+        you can remove the width constraint by providing the <EuiCode>grow</EuiCode> prop.
+      </p>
+    ),
+    props: { EuiText },
+    demo: <TextWidth />,
   }, {
     title: 'Coloring text',
     source: [{