+++ src/components/basic_table/collapsed_item_actions.test.js
@@ -21,7 +21,6 @@ describe('CollapsedItemActions', () => {
           }
         }
       ],
-      visible: true,
       itemId: 'id',
       item: { id: 'xyz' },
       actionEnabled: () => true,