+++ src/components/basic_table/expanded_item_actions.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { DefaultItemAction } from './default_item_action';
 import { CustomItemAction } from './custom_item_action';
 
-export const ExpandedItemActions = ({ actions, visible, itemId, item, actionEnabled }) => {
+export const ExpandedItemActions = ({ actions, itemId, item, actionEnabled, className }) => {
 
   return actions.reduce((tools, action, index) => {
     const available = action.available ? action.available(item) : true;
@@ -16,10 +16,10 @@ export const ExpandedItemActions = ({ actions, visible, itemId, item, actionEnab
       tools.push(
         <CustomItemAction
           key={key}
+          className={className}
           index={index}
           action={action}
           enabled={enabled}
-          visible={visible}
           itemId={itemId}
           item={item}
         />
@@ -28,10 +28,10 @@ export const ExpandedItemActions = ({ actions, visible, itemId, item, actionEnab
       tools.push(
         <DefaultItemAction
           key={key}
+          className={className}
           index={index}
           action={action}
           enabled={enabled}
-          visible={visible}
           itemId={itemId}
           item={item}
         />