+++ src/components/index.js
@@ -58,6 +58,10 @@ export {
 } from './context_menu';
 
 export {
+  EuiDatePicker,
+} from './date_picker';
+
+export {
   EuiDelayHide
 } from './delay_hide';
 
@@ -230,6 +234,7 @@ export {
 } from './spacer';
 
 export {
+  EuiStep,
   EuiSteps,
   EuiSubSteps,
   EuiStepsHorizontal,