+++ src/components/table/table_row_cell.js
@@ -19,6 +19,7 @@ export const EuiTableRowCell = ({
   children,
   className,
   truncateText,
+  showOnHover,
   textOnly,
   colSpan,
   ...rest
@@ -26,20 +27,30 @@ export const EuiTableRowCell = ({
   const contentClasses = classNames('euiTableCellContent', className, {
     'euiTableCellContent--alignRight': align === RIGHT_ALIGNMENT,
     'euiTableCellContent--alignCenter': align === CENTER_ALIGNMENT,
+    'euiTableCellContent--showOnHover': showOnHover,
     'euiTableCellContent--truncateText': truncateText,
     // We're doing this rigamarole instead of creating `euiTableCellContent--textOnly` for BWC
     // purposes for the time-being.
-    'euiTableCellContent--overflowingContent': !textOnly,
+    'euiTableCellContent--overflowingContent': textOnly !== true,
   });
+
+  const childClasses = classNames({
+    'euiTableCellContent__text': textOnly === true,
+    'euiTableCellContent__hoverItem': showOnHover,
+  });
+
+  let modifiedChildren = children;
+
+  if(textOnly === true) {
+    modifiedChildren = <span className={childClasses}>{children}</span>;
+  } else if(React.isValidElement(modifiedChildren)) {
+    modifiedChildren = React.Children.map(children, child => React.cloneElement(child, { className: childClasses }));
+  }
 
   return (
     <td className="euiTableRowCell" colSpan={colSpan}>
       <div className={contentClasses} {...rest}>
-        {
-          textOnly === true
-            ? <span className="euiTableCellContent__text">{children}</span>
-            : children
-        }
+        {modifiedChildren}
       </div>
     </td>
   );
@@ -47,6 +58,7 @@ export const EuiTableRowCell = ({
 
 EuiTableRowCell.propTypes = {
   align: PropTypes.oneOf(ALIGNMENT),
+  showOnHover: PropTypes.bool,
   truncateText: PropTypes.bool,
   children: PropTypes.node,
   className: PropTypes.string,